wip for multithreading
Maybe it's my computer, but it seems like the total time is the same as if I just run it in one go.

Have not figured out how to keep the combinations in sync, so there are duplicates. And the worker seems to dump all the log messages right at the end, so it's harder to monitor.